ddl refactoring
